+++ Client.ts
@@ -1,27 +1,76 @@
 import { IncomingMessage } from "http";
 import { URL } from "url";
 import axios, { AxiosInstance } from "axios";
-import { ClientConfig, AuthenticationResponse, APIError, Hit } from "./types";
+import {
+    ClientConfig,
+    AuthenticationResponse,
+    APIError,
+    Hit,
+    Domain,
+    Filter,
+    Keyword,
+    ScreenClassStats,
+    PlatformStats,
+    CountryStats,
+    BrowserStats,
+    OSStats,
+    ReferrerStats,
+    LanguageStats,
+    VisitorHourStats,
+    ActiveVisitorsData,
+    Growth,
+    ConversionGoal,
+    PageStats,
+    VisitorStats,
+    UTMTermStats,
+    UTMContentStats,
+    UTMCampaignStats,
+    UTMMediumStats, UTMSourceStats, TimeSpentStats
+} from "./types";
 
 const defaultBaseURL = "https://api.pirsch.io";
 const defaultTimeout = 5000;
 const defaultProtocol = "http";
 const authenticationEndpoint = "/api/v1/token";
 const hitEndpoint = "/api/v1/hit";
+const domainEndpoint = "/api/v1/domain";
+const sessionDurationEndpoint = "/api/v1/statistics/duration/session";
+const timeOnPageEndpoint = "/api/v1/statistics/duration/page";
+const utmSourceEndpoint = "/api/v1/statistics/utm/source";
+const utmMediumEndpoint = "/api/v1/statistics/utm/medium";
+const utmCampaignEndpoint = "/api/v1/statistics/utm/campaign";
+const utmContentEndpoint = "/api/v1/statistics/utm/content";
+const utmTermEndpoint = "/api/v1/statistics/utm/term";
+const visitorsEndpoint = "/api/v1/statistics/visitor";
+const pagesEndpoint = "/api/v1/statistics/page";
+const conversionGoalsEndpoint = "/api/v1/statistics/goals";
+const growthRateEndpoint = "/api/v1/statistics/growth";
+const activeVisitorsEndpoint = "/api/v1/statistics/active";
+const timeOfDayEndpoint = "/api/v1/statistics/hours";
+const languageEndpoint = "/api/v1/statistics/language";
+const referrerEndpoint = "/api/v1/statistics/referrer";
+const osEndpoint = "/api/v1/statistics/os";
+const browserEndpoint = "/api/v1/statistics/browser";
+const countryEndpoint = "/api/v1/statistics/country";
+const platformEndpoint = "/api/v1/statistics/platform";
+const screenEndpoint = "/api/v1/statistics/screen";
+const keywordsEndpoint = "/api/v1/statistics/keywords";
 const referrerQueryParams = [
 	"ref",
 	"referer",
 	"referrer",
+    "source",
+    "utm_source"
 ];
 
 /**
  * Client is the Pirsch API client.
  */
 export class Client {
-    private clientID: string;
-    private clientSecret: string;
-    private hostname: string;
-    private protocol: string;
+    private readonly clientID: string;
+    private readonly clientSecret: string;
+    private readonly hostname: string;
+    private readonly protocol: string;
     private client: AxiosInstance;
     private accessToken: string = "";
 
@@ -56,12 +105,12 @@ export class Client {
     }
 
     /**
-     * Hit sends a hit to Pirsch. Make sure you call it in all request handlers you want to track.
+     * hit sends a hit to Pirsch. Make sure you call it in all request handlers you want to track.
      * Also, make sure to filter out unwanted pathnames (like /favicon.ico in your root handler for example).
      * 
      * @param hit all required data for the request.
      * @param retry retry the request in case a 401 (unauthenticated) error is returned. Don't modify this.
-     * @returns an empty promise or an APIError, in case something went wrong
+     * @returns APIError or an empty promise, in case something went wrong
      */
     async hit(hit: Hit, retry: boolean = true): Promise<APIError | null> {
         try {
@@ -80,7 +129,7 @@ export class Client {
             });
             return Promise.resolve<null>(null);
         } catch(e) {
-            if(e.response.status === 401) {
+            if(e.response.status === 401 && retry) {
                 try {
                     await this.refreshToken();
                     return this.hit(hit, false);
@@ -97,7 +146,7 @@ export class Client {
      * hitFromRequest returns the required data to send a hit to Pirsch for a Node request object.
      * 
      * @param req the Node request object from the http package.
-     * @returns a Hit object containing all necessary fields.
+     * @returns Hit object containing all necessary fields.
      */
     hitFromRequest(req: IncomingMessage): Hit {
         const url = new URL(req.url || "", `${this.protocol}://${this.hostname}`);
@@ -111,8 +160,261 @@ export class Client {
             dnt: req.headers["dnt"] as string || "",
             user_agent: req.headers["user-agent"] || "",
             accept_language: req.headers["accept-language"] || "",
-            referrer: this.getReferrer(req, url)
+            referrer: Client.getReferrer(req, url)
         };
+    }
+
+    /**
+     * domain returns the domain for this client.
+     *
+     * @param retry retry the request in case a 401 (unauthenticated) error is returned. Don't modify this.
+     * @returns Domain object for this client.
+     */
+    async domain(retry: boolean = true): Promise<Domain | APIError> {
+        try {
+            const resp = await this.client.get(domainEndpoint, {
+                headers: {
+                    "Content-Type": "application/json",
+                    "Authorization": `Bearer ${this.accessToken}`
+                }
+            });
+
+            if(resp.data.length !== 1) {
+                return Promise.reject<APIError>({
+                    error: ["domain not found"]
+                });
+            }
+
+            return Promise.resolve<Domain>(resp.data);
+        } catch(e) {
+            if(e.response.status === 401 && retry) {
+                try {
+                    await this.refreshToken();
+                    return this.domain(false);
+                } catch(e) {
+                    return e;
+                }
+            }
+
+            return Promise.reject<APIError>(e.response.data);
+        }
+    }
+
+    /**
+     * sessionDuration returns the session duration grouped by day.
+     *
+     * @param filter used to filter the result set.
+     */
+    async sessionDuration(filter: Filter): Promise<TimeSpentStats[] | APIError> {
+        return await this.performGet<TimeSpentStats[]>(sessionDurationEndpoint, filter);
+    }
+
+    /**
+     * timeOnPage returns the time spent on pages.
+     *
+     * @param filter used to filter the result set.
+     */
+    async timeOnPage(filter: Filter): Promise<TimeSpentStats[] | APIError> {
+        return await this.performGet<TimeSpentStats[]>(timeOnPageEndpoint, filter);
+    }
+
+    /**
+     * utmSource returns the utm sources.
+     *
+     * @param filter used to filter the result set.
+     */
+    async utmSource(filter: Filter): Promise<UTMSourceStats[] | APIError> {
+        return await this.performGet<UTMSourceStats[]>(utmSourceEndpoint, filter);
+    }
+
+    /**
+     * utmMedium returns the utm medium.
+     *
+     * @param filter used to filter the result set.
+     */
+    async utmMedium(filter: Filter): Promise<UTMMediumStats[] | APIError> {
+        return await this.performGet<UTMMediumStats[]>(utmMediumEndpoint, filter);
+    }
+
+    /**
+     * utmCampaign returnst he utm campaigns.
+     *
+     * @param filter used to filter the result set.
+     */
+    async utmCampaign(filter: Filter): Promise<UTMCampaignStats[] | APIError> {
+        return await this.performGet<UTMCampaignStats[]>(utmCampaignEndpoint, filter);
+    }
+
+    /**
+     * utmContent returns the utm content.
+     *
+     * @param filter used to filter the result set.
+     */
+    async utmContent(filter: Filter): Promise<UTMContentStats[] | APIError> {
+        return await this.performGet<UTMContentStats[]>(utmContentEndpoint, filter);
+    }
+
+    /**
+     * utmTerm returns the utm term.
+     *
+     * @param filter used to filter the result set.
+     */
+    async utmTerm(filter: Filter): Promise<UTMTermStats[] | APIError> {
+        return await this.performGet<UTMTermStats[]>(utmTermEndpoint, filter);
+    }
+
+    /**
+     * visitors returns the visitor statistics grouped by day.
+     *
+     * @param filter used to filter the result set.
+     */
+    async visitors(filter: Filter): Promise<VisitorStats[] | APIError> {
+        return await this.performGet<VisitorStats[]>(visitorsEndpoint, filter);
+    }
+
+    /**
+     * pages returns the page statistics grouped by page.
+     *
+     * @param filter used to filter the result set.
+     */
+    async pages(filter: Filter): Promise<PageStats[] | APIError> {
+        return await this.performGet<PageStats[]>(pagesEndpoint, filter);
+    }
+
+    /**
+     * conversionGoals returns all conversion goals.
+     *
+     * @param filter used to filter the result set.
+     */
+    async conversionGoals(filter: Filter): Promise<ConversionGoal[] | APIError> {
+        return await this.performGet<ConversionGoal[]>(conversionGoalsEndpoint, filter);
+    }
+
+    /**
+     * growth returns the growth rates for visitors, bounces, ...
+     *
+     * @param filter used to filter the result set.
+     */
+    async growth(filter: Filter): Promise<Growth | APIError> {
+        return await this.performGet<Growth>(growthRateEndpoint, filter);
+    }
+
+    /**
+     * activeVisitors returns the active visitors and what pages they're on.
+     *
+     * @param filter used to filter the result set.
+     */
+    async activeVisitors(filter: Filter): Promise<ActiveVisitorsData | APIError> {
+        return await this.performGet<ActiveVisitorsData>(activeVisitorsEndpoint, filter);
+    }
+
+    /**
+     * timeOfDay returns the number of unique visitors grouped by time of day.
+     *
+     * @param filter used to filter the result set.
+     */
+    async timeOfDay(filter: Filter): Promise<VisitorHourStats[] | APIError> {
+        return await this.performGet<VisitorHourStats[]>(timeOfDayEndpoint, filter);
+    }
+
+    /**
+     * languages returns language statistics.
+     *
+     * @param filter used to filter the result set.
+     */
+    async languages(filter: Filter): Promise<LanguageStats[] | APIError> {
+        return await this.performGet<LanguageStats[]>(languageEndpoint, filter);
+    }
+
+    /**
+     * referrer returns referrer statistics.
+     *
+     * @param filter used to filter the result set.
+     */
+    async referrer(filter: Filter): Promise<ReferrerStats[] | APIError> {
+        return await this.performGet<ReferrerStats[]>(referrerEndpoint, filter);
+    }
+
+    /**
+     * os returns operating system statistics.
+     *
+     * @param filter used to filter the result set.
+     */
+    async os(filter: Filter): Promise<OSStats[] | APIError> {
+        return await this.performGet<OSStats[]>(osEndpoint, filter);
+    }
+
+    /**
+     * browser returns browser statistics.
+     *
+     * @param filter used to filter the result set.
+     */
+    async browser(filter: Filter): Promise<BrowserStats[] | APIError> {
+        return await this.performGet<BrowserStats[]>(browserEndpoint, filter);
+    }
+
+    /**
+     * country returns country statistics.
+     *
+     * @param filter used to filter the result set.
+     */
+    async country(filter: Filter): Promise<CountryStats[] | APIError> {
+        return await this.performGet<CountryStats[]>(countryEndpoint, filter);
+    }
+
+    /**
+     * platform returns the platforms used by visitors.
+     *
+     * @param filter used to filter the result set.
+     */
+    async platform(filter: Filter): Promise<PlatformStats[] | APIError> {
+        return await this.performGet<PlatformStats[]>(platformEndpoint, filter);
+    }
+
+    /**
+     * screen returns the screen classes used by visitors.
+     *
+     * @param filter used to filter the result set.
+     */
+    async screen(filter: Filter): Promise<ScreenClassStats[] | APIError> {
+        return await this.performGet<ScreenClassStats[]>(screenEndpoint, filter);
+    }
+
+    /**
+     * keywords returns the Google keywords, rank, and CTR.
+     *
+     * @param filter used to filter the result set.
+     */
+    async keywords(filter: Filter): Promise<Keyword[] | APIError> {
+        return await this.performGet<Keyword[]>(keywordsEndpoint, filter);
+    }
+
+    private async performGet<T>(url: string, filter: Filter, retry: boolean = true): Promise<T | APIError> {
+        try {
+            if(!this.accessToken && retry) {
+                await this.refreshToken();
+            }
+
+            const resp = await this.client.get(url, {
+                headers: {
+                    "Content-Type": "application/json",
+                    "Authorization": `Bearer ${this.accessToken}`
+                },
+                params: filter
+            });
+            return Promise.resolve<T>(resp.data);
+        } catch(e) {
+            if(e.response.status === 401 && retry) {
+                try {
+                    await this.refreshToken();
+                    return this.performGet<T>(url, filter, false);
+                } catch(e) {
+                    return e;
+                }
+            }
+
+            return Promise.reject<APIError>(e.response.data);
+        }
     }
 
     private async refreshToken(): Promise<APIError | null> {
@@ -129,7 +431,7 @@ export class Client {
         }
     }
 
-    private getReferrer(req: IncomingMessage, url: URL): string {
+    private static getReferrer(req: IncomingMessage, url: URL): string {
         let referrer = req.headers["referer"] || "";
 
         if(referrer === "") {