+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.1.0
+
+* added `source` and `utm_source` to referrers
+* added methods to read statistics
+* updated dependencies
+* fixed retry on 401
+
 ## 1.0.1
 
 * added missing DNT (do not track) header